done likes,and deleting anecdotes

+++ view/likes.php
@@ -7,50 +7,51 @@ $userId = (int) $_POST['user_id'];
 $anecdoteId = (int) $_POST['anecdote_id'];
 $type = $_POST['type'];
 
+if (isset($_SESSION['user'])) {
 // проверяем, голосовал ранее пользователь за эту новость или нет
-$sql = "SELECT count(*) FROM `likes` WHERE `user_id` = $userId AND `anecdote_id` = $anecdoteId";
-$result = $db->prepare($sql);
-$result->execute();
-$qwerty = $result->fetch(PDO::FETCH_ASSOC);
+    $sql = "SELECT count(*) FROM `likes` WHERE `user_id` = $userId AND `anecdote_id` = $anecdoteId";
+    $result = $db->prepare($sql);
+    $result->execute();
+    $qwerty = $result->fetch(PDO::FETCH_ASSOC);
 //var_dump($result);
 //print_r($qwerty['count(*)']);
 //die();
 
 // если что-то пришло из запроса, значит уже голосовал
 //var_dump($result);exit;
-if($qwerty['count(*)'] > 0){
-    if ($type == 'like') $fieldName = 'count_like';
-// делаем запись о том, что пользователь проголосовал
-    $deleteLike = "DELETE FROM `likes` WHERE user_id=$userId AND anecdote_id = $anecdoteId";
-    $result = $db->prepare($deleteLike);
-    $result->execute();
-    $qwerty = $result->fetch();
-
-// делаем запись для новости - увеличиваем количесво голосов(лайк или дизлайк)
-    $deleteCount = "UPDATE `anecdote`.`anecdote` SET `$fieldName`= `$fieldName` - 1 WHERE  `id` = $anecdoteId";
-    $result = $db->prepare($deleteCount);
-    $result->execute();
-    $qwerty = $result->fetch();
-    echo json_encode(array('result' => 'dislike'));
-}else { // если пользователь не голосовал, проголосуем
-
-    if (!$error) {// получем поле для голосования - лайк или дизлайк
+    if ($qwerty['count(*)'] > 0) {
         if ($type == 'like') $fieldName = 'count_like';
 // делаем запись о том, что пользователь проголосовал
-        $addLike = "INSERT INTO `likes` (`user_id`, `anecdote_id`) VALUES ($userId, $anecdoteId)";
-        $result = $db->prepare($addLike);
+        $deleteLike = "DELETE FROM `likes` WHERE user_id=$userId AND anecdote_id = $anecdoteId";
+        $result = $db->prepare($deleteLike);
         $result->execute();
         $qwerty = $result->fetch();
 
 // делаем запись для новости - увеличиваем количесво голосов(лайк или дизлайк)
-        $addCount = "UPDATE `anecdote`.`anecdote` SET `$fieldName`= `$fieldName` + 1 WHERE  `id` = $anecdoteId";
-        $result = $db->prepare($addCount);
+        $deleteCount = "UPDATE `anecdote`.`anecdote` SET `$fieldName`= `$fieldName` - 1 WHERE  `id` = $anecdoteId";
+        $result = $db->prepare($deleteCount);
         $result->execute();
         $qwerty = $result->fetch();
-        echo json_encode(array('result' => 'success'));
+        echo json_encode(array('result' => 'dislike'));
+    } else { // если пользователь не голосовал, проголосуем
+
+        if (!$error) {// получем поле для голосования - лайк или дизлайк
+            if ($type == 'like') $fieldName = 'count_like';
+// делаем запись о том, что пользователь проголосовал
+            $addLike = "INSERT INTO `likes` (`user_id`, `anecdote_id`) VALUES ($userId, $anecdoteId)";
+            $result = $db->prepare($addLike);
+            $result->execute();
+            $qwerty = $result->fetch();
+
+// делаем запись для новости - увеличиваем количесво голосов(лайк или дизлайк)
+            $addCount = "UPDATE `anecdote`.`anecdote` SET `$fieldName`= `$fieldName` + 1 WHERE  `id` = $anecdoteId";
+            $result = $db->prepare($addCount);
+            $result->execute();
+            $qwerty = $result->fetch();
+            echo json_encode(array('result' => 'success'));
+        }
     }
-}
-// делаем ответ для клиента
+    // делаем ответ для клиента
     if ($error) {
 // если есть ошибки то отправляем ошибку и ее текст
         echo json_encode(array('result' => 'error', 'msg' => $error));
@@ -59,3 +60,7 @@ if($qwerty['count(*)'] > 0){
 //// если нет ошибок сообщаем об успехе
 //        echo json_encode(array('result' => 'success'));
 //    }
+} else {
+    echo json_encode(array('msg' =>'Будь ласка залогіньтеся'));
+}
+
